+++ packages/app/public/CHANGELOG.md
@@ -1,7 +1,10 @@
 # Change Log
+
 ### Added
 - New trees for Urgot, Trundle, Malzahar, Yuumi and Teemo
 - New trophies [trophy:damageDealer], [trophy:hardHitter], [trophy:dominatingDamage]
+### Changed
+- Updated to Riot API version 9.11.1.
 
 ## [3.16.0] - 2019-05-21
 ### Added